changed Cloudpoint to double, fixed fusionslam to work with global coordinates

+++ src/main/java/bgu/spl/mics/application/objects/CloudPoint.java
@@ -5,18 +5,18 @@ package bgu.spl.mics.application.objects;
  * These points are used to generate a point cloud representing objects in the environment.
  */
 public class CloudPoint {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public CloudPoint(int x, int y) {
+    public CloudPoint(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     // Getters and Setters
-    public int getX() { return x; }
-    public void setX(int x) { this.x = x; }
+    public double getX() { return x; }
+    public void setX(double x) { this.x = x; }
 
-    public int getY() { return y; }
-    public void setY(int y) { this.y = y; }
+    public double getY() { return y; }
+    public void setY(double y) { this.y = y; }
 }
